+++ crates/solana-utils/build.rs
@@ -6,6 +6,7 @@ fn main() {
         serde: { feature = "serde" },
         client: { feature = "client" },
         anchor: { feature = "anchor" },
+        anchor_lang: {feature = "anchor-lang" },
         client_traits: { feature = "client-traits" },
         http_rpc_sender: { feature = "http-rpc-sender" },
     }